refactor(components) : move <InsertToggleInput /> to new file
